requirements.txt
django>=3.2.5
djangorestframework~=3.12.1